Refactored sqlite-utils insert tests into test_cli_insert.py
